WID-134. Move files regarding HPC execution to a better named folder
