+++ .xircuits/config.ini
@@ -32,40 +32,40 @@ PYUNICORE = DAINT-CSCS
 
 [DAINT-CSCS]
 name =  DAINT-CSCS
-command = python C:/Work/TVB/tvb-ext-xircuits/xai_workflows/pyunicore_config.py
+command = python C:/Work/TVB/tvb-ext-xircuits/hpc_config/pyunicore_config.py
 msg = Submitting job to DAINT-CSCS via pyunicore...
 
 [JUSUF]
 name =  JUSUF
-command = python C:/Work/TVB/tvb-ext-xircuits/xai_workflows/pyunicore_config.py
+command = python C:/Work/TVB/tvb-ext-xircuits/hpc_config/pyunicore_config.py
 msg = Submitting job to JUSUF via pyunicore...
 
 [JUDAC]
 name =  JUDAC
-command = python C:/Work/TVB/tvb-ext-xircuits/xai_workflows/pyunicore_config.py
+command = python C:/Work/TVB/tvb-ext-xircuits/hpc_config/pyunicore_config.py
 msg = Submitting job to JUDAC via pyunicore...
 
 [JURECA]
 name =  JURECA
-command = python C:/Work/TVB/tvb-ext-xircuits/xai_workflows/pyunicore_config.py
+command = python C:/Work/TVB/tvb-ext-xircuits/hpc_config/pyunicore_config.py
 msg = Submitting job to JURECA via pyunicore...
 
 [CINECA-MARCONI]
 name =  CINECA-MARCONI
-command = python C:/Work/TVB/tvb-ext-xircuits/xai_workflows/pyunicore_config.py
+command = python C:/Work/TVB/tvb-ext-xircuits/hpc_config/pyunicore_config.py
 msg = Submitting job to CINECA-MARCONI via pyunicore...
 
 [JUWELS]
 name =  JUWELS
-command = python C:/Work/TVB/tvb-ext-xircuits/xai_workflows/pyunicore_config.pyy
+command = python C:/Work/TVB/tvb-ext-xircuits/hpc_config/pyunicore_config.pyy
 msg = Submitting job to JUWELS via pyunicore...
 
 [CINECA-M100]
 name =  CINECA-M100
-command = python C:/Work/TVB/tvb-ext-xircuits/xai_workflows/pyunicore_config.py
+command = python C:/Work/TVB/tvb-ext-xircuits/hpc_config/pyunicore_config.py
 msg = Submitting job to CINECA-M100 via pyunicore...
 
 [CINECA-G100]
 name =  CINECA-G100
-command = python C:/Work/TVB/tvb-ext-xircuits/xai_workflows/pyunicore_config.py
+command = python C:/Work/TVB/tvb-ext-xircuits/hpc_config/pyunicore_config.py
 msg = Submitting job to CINECA-G100 via pyunicore...
